requirements.txt
faiss-cpu
sentence-transformers
pandas
numpy
requests
beautifulsoup4
fastapi
uvicorn
streamlit